requirements.txt
amqp==5.3.1
asgiref==3.10.0
attrs==25.4.0
autobahn==25.10.2
Automat==25.4.16
billiard==4.2.2
boto3==1.40.71
botocore==1.40.71
celery==5.5.3
certifi==2025.11.12
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
dj-rest-auth==7.0.1
Django==5.2.8
django-anymail==13.1
django-cors-headers==4.9.0
django-redis==6.0.0
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.11
incremental==24.7.2
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
msgpack==1.1.2
packaging==25.0
pillow==12.0.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
redis==7.0.1
referencing==0.37.0
requests==2.32.5
rpds-py==0.28.0
ruff==0.14.4
s3transfer==0.14.0
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
zope.interface==8.1
